+++ build.gradle
@@ -30,13 +30,8 @@ dependencies {
 	runtimeOnly 'mysql:mysql-connector-java'
 	annotationProcessor 'org.projectlombok:lombok'
 	testImplementation 'org.springframework.boot:spring-boot-starter-test'
-
 	// https://mvnrepository.com/artifact/org.jsoup/jsoup
 	implementation group: 'org.jsoup', name: 'jsoup', version: '1.15.3'
-
-	// https://mvnrepository.com/artifact/org.springframework.boot/spring-boot-starter-webflux
-	implementation group: 'org.springframework.boot', name: 'spring-boot-starter-webflux', version: '2.7.4'
-
 	//openFeign
 	implementation 'org.springframework.cloud:spring-cloud-starter-openfeign'
 }
