+++ src/main/java/com/shawn/notification/Service/Collector.java
@@ -1,4 +1,4 @@
-package com.shawn.notification;
+package com.shawn.notification.Service;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
