+++ src/main/java/com/shawn/notification/domain/SlackClient.java
@@ -1,6 +1,7 @@
-package com.shawn.notification;
+package com.shawn.notification.domain;
 
 import com.shawn.notification.config.OpenFeignSlackConfig;
+import com.shawn.notification.dto.SlackMessageRequestDto;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
