+++ src/test/java/com/shawn/notification/SlackClientServiceTest.java
@@ -1,5 +1,7 @@
 package com.shawn.notification;
 
+import com.shawn.notification.Service.SlackClientService;
+import com.shawn.notification.dto.SlackMessageRequestDto;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
